backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# GraphQL
strawberry-graphql[fastapi]==0.216.1

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# HTTP Client
httpx==0.25.2

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
ruff==0.1.6
isort==5.12.0
